internalise data (fixes #9)
This creates internal objects of the value set data, removing the
need for calls to `get()`. The package should now be usable as a
dependent package and without the need to call `library()` (i.e.
via `::`). I've not removed any of the data in data\ nor altered
how it is documented in R/data.R