Only use internal (but exported data)
This commit deletes the data folder but creates csv for the raw
valuesets in the data-raw folder. The data-raw also contains a
small script which facilitates easy updating of the internal
R/sysdata.rda file. All valuesets and ranges are exported
for users to access but STATES and REFERENCES are not.