WIP - References (#24)
* remove whitespace

* Update TTO references within R/data.R

This commit adds refererences for:
Argentina
Denmark
France
Germany
Italy
Japan
Netherlands
Spain
South Korea
UK
USA
Zimbabe

* Add valuesets.csv

* update TTO references in valuesets.csv

Add missing TTO references for France, Netherlands, South Korea
and Spain.

* whitespace changes

* Fix ordering of sources in VT documentation

* Add india VT reference

* add VAS references where possible

It seems a lot of the VAS references do not have DOI, so will
probably need to link to the comparative review in these cases.

* Standardize formatting and fix minor typos

* rerender docs

* mv valuesets.csv to data-raw/

* Only use internal (but exported data)

This commit deletes the data folder but creates csv for the raw
valuesets in the data-raw folder. The data-raw also contains a
small script which facilitates easy updating of the internal
R/sysdata.rda file. All valuesets and ranges are exported
for users to access but STATES and REFERENCES are not.

* Add references to return output of valuesets

* Update DESCRIPTION

Updated to match current dev release version.

Co-authored-by: fragla <fragla@users.noreply.github.com>